LPS-175508: Move code to portal-search
