+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/configuration/AsahSearchKeywordsConfiguration.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.internal.configuration;
+
+import aQute.bnd.annotation.metatype.Meta;
+
+import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
+
+/**
+ * @author Petteri Karttunen
+ */
+@ExtendedObjectClassDefinition(
+	category = "search", scope = ExtendedObjectClassDefinition.Scope.COMPANY
+)
+@Meta.OCD(
+	id = "com.liferay.portal.search.internal.configuration.AsahSearchKeywordsConfiguration",
+	localization = "content/Language",
+	name = "asah-search-keywords-configuration-name"
+)
+public interface AsahSearchKeywordsConfiguration {
+
+	@Meta.AD(deflt = "14400", name = "cache-timeout", required = false)
+	public int cacheTimeout();
+
+}
