+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/suggestions/spi/AsahRecentSearchKeywordsSuggestionsContributor.java
@@ -1,18 +1,18 @@
 /**
  * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
  *
- *
- *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
  */
 
-package com.liferay.search.experiences.internal.suggestions.spi;
+package com.liferay.portal.search.internal.suggestions.spi;
 
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
@@ -30,7 +30,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Petteri Karttunen
  */
 @Component(
-	configurationPid = "com.liferay.search.experiences.internal.configuration.AsahSearchKeywordsConfiguration",
+	configurationPid = "com.liferay.portal.search.internal.configuration.AsahSearchKeywordsConfiguration",
 	enabled = false,
 	property = "search.suggestions.contributor.name=asahRecentSearchKeywords",
 	service = SuggestionsContributor.class
