+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/web/cache/AsahSearchKeywordsWebCacheItem.java
@@ -1,18 +1,18 @@
 /**
  * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
  *
- *
- *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
  */
 
-package com.liferay.search.experiences.internal.web.cache;
+package com.liferay.portal.search.internal.web.cache;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
@@ -27,9 +27,9 @@ import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.webcache.WebCacheItem;
 import com.liferay.portal.kernel.webcache.WebCachePoolUtil;
+import com.liferay.portal.search.internal.configuration.AsahSearchKeywordsConfiguration;
+import com.liferay.portal.search.internal.util.AsahUtil;
 import com.liferay.search.experiences.blueprint.exception.InvalidWebCacheItemException;
-import com.liferay.search.experiences.internal.configuration.AsahSearchKeywordsConfiguration;
-import com.liferay.search.experiences.internal.util.AsahUtil;
 
 import java.net.HttpURLConnection;
