--- modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/configuration/AsahSearchKeywordsConfiguration.java
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.search.experiences.internal.configuration;
-
-import aQute.bnd.annotation.metatype.Meta;
-
-import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
-
-/**
- * @author Petteri Karttunen
- */
-@ExtendedObjectClassDefinition(
-	category = "search-experiences",
-	scope = ExtendedObjectClassDefinition.Scope.COMPANY
-)
-@Meta.OCD(
-	id = "com.liferay.search.experiences.internal.configuration.AsahSearchKeywordsConfiguration",
-	localization = "content/Language",
-	name = "asah-search-keywords-configuration-name"
-)
-public interface AsahSearchKeywordsConfiguration {
-
-	@Meta.AD(deflt = "14400", name = "cache-timeout", required = false)
-	public int cacheTimeout();
-
-}
